pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyAFL"
version = "0.1.0"
description = "AFL match + player stats scraper"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Sam Vaughan" }]
dependencies = [
  "requests",
  "beautifulsoup4",
  "pandas",
  "pyarrow",
  "tqdm",
  "duckdb",
  "requests-cache>=1.2.1",
  "ipython>=9.6.0",
  "ipdb>=0.13.13",
  "lxml>=6.0.2",
  "matplotlib>=3.10.7",
  "snowflake-connector-python>=4.0.0",
  "scipy>=1.16.2",
]

[tool.setuptools.packages.find]
where = ["src"]   # your package lives at src/nrl_scraper
